Secret Function for HHil
A Secret encapsulates a function that accepts a single integer
parameter and returns an integer.
Secret is a given but you don’t know its implementation. In Java or a
language of your choice,
write an application that determines if Secret is an additive function
secret(x+y) = secret(x) +
secret(y) for all combinations of prime numbers less than N where N is
also a given.
Assume you are going to ship this code. Use your judgement to determine
what you need to
include. Document any assumptions or choices you make in your
implementation. Also, include
an explanation of your choice of language.